# backend/eVotingApi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Online E-Voting Prototype",
    "version": "v1"
  },
  "paths": {
...
    "/api/Elections": {
      "get": {
        "tags": [
          "Elections"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Election"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Election"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Election"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Elections"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Election"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Election"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Election"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Elections/{id}": {
      "get": {
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Election"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Election"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Election"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Voters/{id}": {
      "get": {
        "tags": [
          "Voters"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Votes/{id}": {
      "get": {
        "tags": [
          "Votes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Election": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "integer",
            "format": "int64"
          },
          "startTime": {
            "type": "integer",
            "format": "int64"
          },
          "endTime": {
            "type": "integer",
            "format": "int64"
          },
          "electionTitle": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}